First check <Peaks> tag. (We might want  to override the file's units or label for our purposes, and we don't want to mess with the file records.)
Then check ##XLABEL/##YLABEL -- this is what they are meant for, and we were not treating those properly before -- instead, we were saving these values as xUnits and yUnits, respectively, if they were found in a file. (None of our sample files have ##XLABEL or ##YLABEL, though, so we never noticed this bug.)

Finally, check ##XUNITS/##YUNITS.

In any case, never return "ARBITRARY UNITS" and instead return "" in that case. This is unambiguous.

git-svn-id: https://svn.code.sf.net/p/jspecview/svn@871 1262bc75-0126-4021-b206-820d13540e36