Updated version of the to do list:
-when in the split overlay version: addition of two boxes next to the
one to toggle between stacked and split, one to clear the selected
spectrum (X), and one to minimize it only (-)

-a save "view" (overlay) option: writes a jdx file that will be able
to call up changes to a current overlay similar to "save as," so that
if one of the used spectra is used in a different file, separate
changes and analyses can be made without altering the saved view

-both jvalues and integration should have the option to
show/hide/reset: show would pull up the most recently hidden
measurements, reset would clear the measurements for that "view"

-in the jsvtest.htm page, when there is a CNMR and IR overlay shown
nothing happens when a proton is clicked on a molecule in the Jmol
applet. The viewer should have the option to either add H-NMR with
current overlay or to load only H-NMR (in this case the viewer could
have already saved the C-NMR and IR changes as a "view" to return to
later--could be done through a pop up box

-when you offset a spectrum the peaks are able to go off the page, the
user cannot click on the parts of the spectra that are off the grid

- when the spectra has coupling constants, when it is zoomed to view a
spot that does not have the constants, the constants still show up
even if their peaks are off the page (they are where the peaks would
be if the whole spectrum were blown up)

-Integration:
 --changing the command from "integrate" to "integration"
   ---also what is the difference between integrate on/off and
showintegration true/false/toggle?
 --in the right-click file box there is no show integration
option---might not be necessary, but would be nice for the time being
due to the mac bug (disables keyboard function when type into the
right-click script box) we found today making it impossible to operate
the integration

git-svn-id: https://svn.code.sf.net/p/jspecview/svn@998 1262bc75-0126-4021-b206-820d13540e36